Transmute rule -214 into a mutable rule.
I intend to make it possible to amend more than one rule at a time. To do so, the rule allowing rule mutation must be modified, but that rule is currently immutable.

This change satisfies rule -214: it transmutes a single rule from immutable to mutable.

This change satisfies rule -209: the pull request implementing this rule has been numbered by Github.

This change satisfies rule -208: the nature of the change is explicit in the title of this proposal.

This change satisfies rule -211: the rules-change proposed is written down.

This change satisfies rule -104: there will be 13 mutable rules.

All other rules are trivially satisfied.